dashboard/backend/sql_client.py
import pyodbc
import os
from dotenv import load_dotenv

load_dotenv()

def get_sql_conn():
    conn_str = (
        f"DRIVER={{ODBC Driver 17 for SQL Server}};"
        f"SERVER={os.getenv('SQL_SERVER')};"
        f"DATABASE={os.getenv('SQL_DB')};"
        f"UID={os.getenv('SQL_USER')};"
        f"PWD={os.getenv('SQL_PASSWORD')}"
    )
    return pyodbc.connect(conn_str)
