dashboard/backend/requirements.txt
fastapi
uvicorn
redis
pyodbc
python-dotenv
